chore: separação de arquivo node.hpp, alteração de diretório
