+++ program/includes/Editor/Windows/UINodeEditor/Interfaces/OnUpdateNode.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Editor/Windows/UINodeEditor/Nodes/Node.hpp"
+#include "Editor/Windows/UINodeEditor/Node/Node.hpp"
 #include "Editor/Windows/UINodeEditor.hpp"
 
 #include <ImGuiNodeEditor/imgui_node_editor.h>
